+++ providers/darwin/memory_darwin.go
@@ -15,25 +15,21 @@
 // specific language governing permissions and limitations
 // under the License.
 
-// +build darwin,amd64,cgo
+// +build amd64,cgo arm64,cgo
 
 package darwin
 
 import (
-	"syscall"
-	"time"
-
 	"github.com/pkg/errors"
 )
 
-const kernBoottimeMIB = "kern.boottime"
+const hwMemsizeMIB = "hw.memsize"
 
-func BootTime() (time.Time, error) {
-	var tv syscall.Timeval
-	if err := sysctlByName(kernBoottimeMIB, &tv); err != nil {
-		return time.Time{}, errors.Wrap(err, "failed to get host uptime")
+func MemTotal() (uint64, error) {
+	var size uint64
+	if err := sysctlByName(hwMemsizeMIB, &size); err != nil {
+		return 0, errors.Wrap(err, "failed to get mem total")
 	}
 
-	bootTime := time.Unix(int64(tv.Sec), int64(tv.Usec)*int64(time.Microsecond))
-	return bootTime, nil
+	return size, nil
 }
